GraphPanel extends ContentView
git-svn-id: https://wush.net/svn/mindalliance@877 baad322d-9929-0410-88f0-f92e8ff3e1bd

+++ playbook/trunk/src/main/groovy/com/mindalliance/channels/playbook/graph/svg/SVGTranslate.groovy
@@ -11,8 +11,8 @@ import com.mindalliance.channels.playbook.graph.svg.SVGTransform
  */
 class SVGTranslate extends SVGTransform {
 
-    int x = 0
-    int y = 0
+    double x = 0.0
+    double y = 0.0
 
     boolean isTranslate() {
         return true
@@ -32,7 +32,7 @@ class SVGTranslate extends SVGTransform {
     }
 
     boolean isIdentity() {
-        return x == 0 && y == 0
+        return x == 0.0 && y == 0.0
     }
 
 }
